refactor: Se modifica modulos con implementaciones incorrectas

+++ Backend/src/main/java/com/sacm/Backend/Case/Users/Users_Patients/Infrastructure/persistence/Repositories/SpringDataUserRepository.java
@@ -3,5 +3,5 @@ package com.sacm.Backend.Case.Users.Users_Patients.Infrastructure.persistence.Re
 import com.sacm.Backend.Case.Users.Users_Patients.Infrastructure.persistence.Entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpringDataUserRepository extends JpaRepository<UserEntity, Integer> {
+public interface SpringDataUserRepository extends JpaRepository<UserEntity, Long> {
 }
